Rename file to prevent exception when running db:setup task.
