+++ app/models/document_factory_registry.rb
@@ -11,7 +11,7 @@ require "validators/manual_validator"
 require "validators/medical_safety_alert_validator"
 require "validators/null_validator"
 require "validators/raib_report_validator"
-require "validators/vehicle_recalls_and_faults_validator"
+require "validators/vehicle_recalls_and_faults_alert_validator"
 
 require "builders/manual_document_builder"
 require "manual_with_documents"
